split consoleproc list management and APIs into separate source
